Move run diagnostics to internal type module
